refactor gulpfile as it was getting big. dispatch each task in a specific file in gulp/task delete test task as it does not work with rollup
