+++ package.json
@@ -10,8 +10,7 @@
     "build": "rollup -c -o dist/bundle.js && babel dist/bundle.js | uglifyjs -cm | indexhtmlify --title js13kgame --style src/main.css | html-minifier --minify-css > dist/index.html",
     "gulp-prod": "gulp --production",
     "gulp-prod-debug": "gulp --production --sourcemaps",
-    "gulp-dev": "gulp",
-    "gulp-watch": "gulp watch",
+    "gulp-dev": "gulp watch",
     "gulp-release": "gulp release",
     "release": "gzip -9 -c dist/index.html > dist/game.gz",
     "deploy": "npm run build && npm run release && gh-pages -d ./dist -m \"Size: $(find ./dist -maxdepth 1 -printf '%s\t%f\n' | grep game.gz | grep -Po '[0-9]+')\""
@@ -51,6 +50,7 @@
     "jasmine-reporters": "^2.2.0",
     "jshint": "^2.9.3",
     "jshint-stylish": "^2.2.1",
+    "require-dir": "^0.3.0",
     "rollup": "0.34.10",
     "rollup-watch": "2.5.0",
     "uglify-js": "2.7.3",
